requirements.txt
rdkit-pypi
pandas
numpy
scipy